requirements.txt
# requirements.txt
streamlit
google-generativeai
pymupdf